requirements.txt
beautifulsoup4==4.6.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-filter==2.1.0
django-modelcluster==4.3
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.1
draftjs-exporter==2.1.5
feedparser==5.2.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
Markdown==3.1
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
requests==2.21.0
six==1.12.0
Unidecode==1.0.23
urllib3==1.24.1
wagtail==2.4
webencodings==0.5.1
Willow==1.1
